fix!: Hide the internal representation of EntryMode
Change the gix-object API so that client code can't explicitely rely on
the internal representation of `EntryMode`.

This is necessary as we need to change it to allow round-trip behaviour
for modes like `b"040000"` and `b"40000"` which currently share the
  `0o40000u16` representation.

Use the opportunity to sprinkle a couple of optimizations in the parsing
of the EntryMode since we had to go deep in understanding this code
anyway, so may as well. Mostly, don't `reverse` the bytes when parsing.

```
TreeRefIter()           time:   [46.251 ns 46.390 ns 46.549 ns]
                        change: [-17.685% -16.512% -15.664%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
```

Also add a test that shows the current incorrect behaviour.