Instrument make_remote_repos.sh to view `config` corruption
The change in the previous commit of switching to the non-shim
`bash.exe` in `(git root)/usr/bin` causes problems, because the
environment may not be correct for shell commands and scripts.
In particular, the `PATH` might not enable standard POSIX tools to
be found, or the tools that are found may interoperate incorrectly
with the shell. The latter caused failures in #1862 in an analogous
choice of `sh.exe`, which were addressed by preferring the shim
when available. See:

- https://github.com/GitoxideLabs/gitoxide/pull/1862#issuecomment-2692158831

Here, 90 tests started to fail when the test suite was run locally
from PowerShell (i.e. not a Git Bash environment) on a Windows 10
system that, in addition to a full Git for Windows installation,
contains a separate non-GfW MSYS2 installation whose `bin`
directories are in `PATH` even in non-MSYS2 environments. The
failures were described, and most of them investigated, as follows:

- https://github.com/GitoxideLabs/gitoxide/pull/1864#issuecomment-2724464883
- https://gist.github.com/EliahKagan/3c5eebd091e66d8c912fddbce0a064cd
- https://gist.github.com/EliahKagan/17066ad1f7b0aa98e4fdf4642abe1d93

Most failures, including all those that were unintuitive, were
directly or indirectly due to the `make_remote_repos.sh` fixture
script encountering the error:

    fatal: bad config line 10 in file ./config

This happened due to the same incorrect behavior of `>>`, when used
by a shell that links to one `msys-2.0.dll` running a program that
links to a different `msys-2.0.dll` of another version or build, as
caused the failure encountered with the non-shim in #1862.

(It may be the handful of other failures are also caused by this
`>>` problem, but as of now that has not been examined.)

This commit temporarily instruments that fixture script so that,
when tests are run, the observations and analysis in the last gist
above can be confirmed. (These changes are also shown there.)