feat: Look for bash in `(git root)/bin`, then `(git root)/usr/bin`
This changes `bash_program()` so that it will find the `bash.exe`
provided by Git for Windows that is most reliable for our use in
runinng test fixture scripts, of those that are available. First
it uses the shim, but falls back to the non-shim if the shim is
not available. If neither is found, then the fallback of using the
simple command `bash.exe` (which triggers a path search when run)
continues to be used.