feat: Document `gix_testtools::bash_program()` and make it public
To make it easier for users of `gix-testtools` to diagnose problems
or verify that the fallback for running fixutre scripts without
usable shebangs (which is effectively how any fixture shell script
is run on Windows), the formerly private `bash_program()` is now
public.

However, it is not recommend to rely on this specific value or on
observed behavior of how it is computed. The details may change at
any time.

The purpose of `bash_program()` and how it is used internally by
`gix-testtools` is also now documented explicitly. Broad details of
how it searches or guesses what path to use are likewise documented,
with a caveat that changes to them are not considered breaking.