Temporarily repeat the fs snapshot journey many times
This is to smoke out nondeterministic failures on any platforms
that have them, so as to gain greater confidence that the
forthcoming fix is effective.

The two major matrix job definitions in `ci.yml` are accordingly
made so a failing job does not cancel the other jobs, so all jobs
that would fail can get as far as to show their failures. This is
likewise temporary.