Drop trailing `,` just before `)` on same line in function calls
These appear unintentional, and in most cases may have been
introduced before rustfmt collapsed the lines, or with the idea
that they would cause rustfmt to expand the lines but where that
did not occur, or in an earlier version of the code where the last
argument was longer and thus split onto another line.

This does not remove a trailing `,` after the same argument when
the argument is on its own line. Although this does remove trailing
`,` in some macro calls (e.g., `format!`), this of course avoids
making such changes in macro definitions, where `,)` is semantic.