Add regenerated `generate_git_date_baseline` archive
This was regenerated via `cargo nextest --workspace --no-fail-fast`
and verifying that there were no failures, then committing the
one changed file, on an Arch Linux system with Git 2.49.0.

This is the only tracked archive that changes due to the script
changes in the previous commit. The other changed scripts are
either to scripts whose generated archives are `.gitignore`d, or
to scripts that are not fixture scripts (so no archives at all).