Add missing `apt-get update` in CI `test` job
The CI `test` job's "Setup dependencies" step had `apt-get install`
without having previously run `apt-get update`. This often does not
work, and in a CI environment one should not typically expect it to
work because package indexes in a virtual machine just provisioned
from an image, if present at all, may be very outdated. But for
some reason it had been working until recently, when breakages were
observed, including in #1924 (though the breakage is not related to
the changes in that PR).

This runs `apt-get update` before `apt-get install` in that CI job,
as had already been done in the other CI jobs and as had likely
always been intended. This should make the "Setup dependencies"
step work again.