Delist LicenseRef-ring, as ring no longer uses it
Before this change, we get a warning:

    $ cargo deny --workspace --all-features check licenses
    warning[license-not-encountered]: license was not encountered
       ┌─ /home/ek/source/repos/gitoxide/deny.toml:32:6
       │
    32 │     "LicenseRef-ring",
       │      ━━━━━━━━━━━━━━━ unmatched license allowance

    licenses ok

The same warning is shown in the `cargo-deny` job check on CI.

This happens because `ring` no longer uses a custom/nonstandard
license, instead using `Apache-2.0 AND ISC` since version 0.17.10.
(See https://github.com/briansmith/ring/pull/2402 and
https://crates.io/crates/ring/versions for details.)

Nothing in this workspace depends directly or directly on old
versions of `ring` that use that license, so this removes it from
the list of licenses in `deny.toml`.