revset: don't crash if revset engine returns non-BackendError
We call `RevsetEvaluationError::expect_backend_error()` in lots of
places. If the revset engine had returned a non-`BackendError`, it
would result in a panic. For example, a revset engine might return
something like `RevsetEvaluationError::Other("not supported".into())`
for some revsets. We shouldn't crash when this happens.

We may want to create some higher-level `RepoError` to return here,
but for now let's just avoid the crash.