config-schema: rename schema test module used by datatest
Because the datatests don't use the default libtest harness, datatests
and normal tests cannot be mixed in the same module or even binary.

`test_config_schema` is renamed, both to reflect that it is unlike the
other `test_*` modules, but also to make the name available for "normal"
tests related to the config schema.